refactor: Structure by componenets instead of technical role
