+++ src/components/Event/model/Event.ts
