+++ src/components/Topic/model/Topic.ts
